+++ test/test_yawast.rb
@@ -14,4 +14,54 @@ class TestYawast < Minitest::Test
 
     restore_stdout
   end
+
+  def test_non_www_redirect
+    override_stdout
+
+    original = Yawast::Shared::Uri.extract_uri'https://www.adamcaudill.com'
+    new = Yawast::Scanner::Core.check_www_redirect original.copy
+
+    assert original.host != new.host, "Host not changed: '#{new}'"
+    assert stdout_value.include?('Non-WWW redirect to'), "Non-WWW Redirect not found in: #{stdout_value}"
+
+    restore_stdout
+  end
+
+  def test_www_redirect
+    override_stdout
+
+    original = Yawast::Shared::Uri.extract_uri'https://apple.com'
+    new = Yawast::Scanner::Core.check_www_redirect original.copy
+
+    assert original.host != new.host, "Host not changed: '#{new}'"
+    assert stdout_value.include?('WWW redirect to'), "WWW Redirect not found in: #{stdout_value}"
+
+    restore_stdout
+  end
+
+  def test_no_redirect
+    override_stdout
+
+    original = Yawast::Shared::Uri.extract_uri'https://adamcaudill.com'
+    new = Yawast::Scanner::Core.check_www_redirect original.copy
+
+    assert original.host == new.host, "Host changed: '#{new}'"
+    assert !stdout_value.include?('Non-WWW redirect to'), "Non-WWW Redirect found in: #{stdout_value}"
+    assert !stdout_value.include?('WWW redirect to'), "WWW Redirect found in: #{stdout_value}"
+
+    restore_stdout
+  end
+
+  def test_non_www_redirect_scheme
+    override_stdout
+
+    original = Yawast::Shared::Uri.extract_uri'http://apple.com'
+    new = Yawast::Scanner::Core.check_www_redirect original.copy
+
+    assert original.host != new.host, "Host not changed: '#{new}'"
+    assert stdout_value.include?('WWW redirect to'), "WWW Redirect not found in: #{stdout_value}"
+    assert original.scheme != new.scheme, "Scheme not changed: Original: '#{original}' - New: '#{new}'"
+
+    restore_stdout
+  end
 end