+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.8.0 - In Development
 
+* [#170](https://github.com/adamcaudill/yawast/issues/170) - When checking HEAD, follow redirects
 * [#169](https://github.com/adamcaudill/yawast/issues/169) - Bug: Error in connecting to SSL Labs
 
 ## 0.7.1 - 2019-05-07